Avoid sending PacketOut to the 'OFPP_LOCAL' port.
OFPP_LOCAL a virtual port that the switch uses to connect to the
controller. Thus, we cannot send a 'PacketOut' with an ActionOutput to
that port. If we do so, we are going to disconnect ourselves from the
switch.

Until now we were just checking for the 'OFPP_LOCAL' sized to v0x01,
which was an UBInt16 (65534). Now we are also checking it to the UBInt32
from v0x04. And instead of using the values, we are using the enums.

FIX kytos/of_lldp#5

Signed-off-by: Carlos Eduardo M. Santos <cadu AT ncc DOT unesp DOT br>